+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next Release
 
++ [**IMPROVEMENT**] src/magic.cpp: Rename flag_converter to flags_converter.
 + [**FEATURE**] inc/magic.hpp, src/magic.cpp: Make setting parameters of magic using parameter_value_map_t possible.
 + [**FEATURE**] inc/magic.hpp, src/magic.cpp: Make setting flags of magic using flags_container_t possible.
 + [**FEATURE**] inc/magic.hpp, src/magic.cpp: Make construction of magic using the flags_container_t possible.